Fix non-deterministic failures in test_cross_element_names (#18)
This test was failing in CI sometimes, and began to fail quite persistently,
due to an issue with the way we were generating invalid models.

The core issue is that the invalid model element is being generated by
mutating an existing valid model to conform to the error. For cross-element
name checks we had to rename a dimension, but if you rename a primary
time dimension the model validator fails with the wrong error.

To sidestep this problem for the time being we simply restrict the mutated
dimension to one of type CATEGORICAL. Longer term we should create very simple
invalid models that match the validation error types, or models with parameterized
names so we can pass concrete values to the template to generate the errors we need.