Adding Validation and Test for Composite Identifiers
MetricFlow enforces that each identifier is defined uniquely, one aspect of this is that the sub identifiers are the same.
This is not a problem when an identifier is only requested once in a model. But this can become a problem when a model requests the same identifier across multiple data sources of a model. Each time that identifier is requested, it must share the same sub-identifier structure.