Removing `time_format` from `DimensionTypeParams` call-site because redundant
The `DimensionTypeParams` defaults to `YYYY-MM-DD` so including the `time_format` set to this value in the call-site is redundant. Taking the time in this PR to clean up the call-site in this validation. Reason it was there in the port of this validation is because we needed it in Legacy. "If you've got the broom out, clean the floor."