Use SourceNodeBuilder to provide inputs to DataflowPlanBuilder.
The source node builder allows the data flow plan builder build a join plan
using nodes that have the plot time dimension available. Previously, it
created read nodes that provided access to only the elements defined in the
data sources, so the source node builder creates a set of input nodes that
include the plot time dimension so that it can be used for query resolution.

Another change that this includes is updating the constructor for the dataflow
plan builder to include the node output resolver and the column association
resolver so that those objects can be instantiated once in the MF engine and
used everywhere. This helps ensure consistency (e.g. the same column name
resolution happens in all objects).