Make validate-configs print success when there are no issues
Previously the validate-configs command only printed the success
message if the 'build_result.issues' was 'None'. However, the
current implementation of the ModelValidator returns an empty tuple
for the issues when there are no issues. Thus the validate-configs
command never printed the success message. This change checks the
boolean value of 'build_result.issues' which evaluates to 'False'
for 'None' and empty sequences.