Add 'dry_run' method to sql_client and plumb through abstract implementations
Sometimes we don't want to actually query the a given statement
against the data warehouse, but just want to check that the query
will work with the data warehouse, i.e. a dry run. The most
straight forward way to do this is by making the query statement
into an 'EXPLAIN' sql statement.