Use '.dry_run' instead of '.query' for DataWarehouseModelValidator
The actual result of the statement being run against the data
warehouse during Data Warehouse Validations is not needed. We are
only checking that the statement is considered valid by the data
warehouse. Additionally, '.dry_run' is a less intensive process
than '.query'. Thus '.dry_run' is used instead of '.query'.