Add data warehouse data source task builder and use in the validation runner
This probably could have been two commits, but I think this one
falls on the cusp. Anyway, this commit adds two things. First, it
adds a DataWarehouseTaskBuilder with some generalized private
methods for building tasks and we use it here to generate tasks
for validating the data source defintions of the user configured
model. Then we added helper, validate_data_sources, to the
DataWarehouseModelValidator which simply generates the tasks and
then runs them through the task runner.