Run data warehouse validations if build has warnings but no errors
This probably could have been two commits, but again here we are
with just one. Things were to mixed to divide the work properly.
The big thing is prior to adding DW validations breaking when there
were model build result issues was unnecessary. However, now if
there are issues build result issues for the model, we don't want
to run the data warehouse validations. Except this is only true if
there are issues of level FATAL or ERROR, whereas WARNING and
FUTURE_ERROR are okay. So now we parse what issue levels we have
and do control flow accordingling.

However, having all this extra control flow means that we were
repeating the code for printing validation issues a ton. To keep
things DRY, a helper has been added to take care of printing a
given list of issues