Refactor validate-configs cli command to abstract data warehouse validations
For each data warehouse model validation type we ran we were
writing a fair bit of duplicate code. Although with only two
types of data warehouse model validations running this isn't
overbearing, we plan on soon having 5 with the addition of
dimensions, measures, and identifiers. Thus it makes sense to
abstract this duplicate code to keep things nice and dry.