Add on_fail_subtasks to DataWarehouseValidationTask
Sometimes we'll want to check sub parts of a query if a large query
failes. For instance to efficiently check that all the dimensions
on a data source exist in the data warehouse, we query them all in
one query. If it passes, all of them are good. If it fails, then
one or more of the dimensions have an issue. However, the big query
will only tell us the 'first' dimension that has an issue. So we'll
want to iterate through all the dimensions individually.