Temporarily disable linter warnings
Linter warnings (and warnings in general) can be very noisy. At
this point it's most important simply to identify the yaml issues
which will cause issues when we parse it. The warnings, which are
things that won't break parsing, end up creating A LOT of noise.
Thus for the time being, until we have a better way to surface
warnings in the CLI, we won't raise linter warnings.