Refactor to ValidationIssue classes to use pydantic BaseModel
MetricFlow is used in other projects. Sometimes MetricFlow data
might need to be returned over an API. For ValidationIssues to be
returnable they need to be jsonify-able. By making the
ValidationIssue classes use the pydantic BaseModel class, we get
jsonability built in as well as a number of other helper methods.