Implement and use ValidationContext instead of model_object_reference for ValidationIssues
Historically we have used model_object_reference (an OrderedDict) to
give context to ValidationIssues. However, this has some drawbacks.
First the OrderedDict could be over specified (i.e. what if both a
dimension and a measure name are specified?). Second it's not easily
JSON serializable. Thus we've moved to something more concrete,
ValidationContext and sub-classes thereof. This new class, and sub-
classes, provides a hierarchy of contexts which all provide the same
interface for generating human readable context strings.
Additionally, ValidationContexts are of our specifically JSON
serializable 'BaseModel' class. The one hickup is that in most
cases we want to generate the ValidaitonContext from an available
'Metadata' object for 'DataSources', 'Metrics', and
'Materializations'. It'd be useful at some point to guarantee the
'Metadata' object on these objects to simply the code somewhat.