Add tests for metric config parsing
The core MetricFlow model parser converts user-defined YAML files
into MetricFlow model objects via a combination of standard,
jsonschema-backed YAML parsing and Pydantic BaseModel object
parsing. This is managed by a set of custom parsers.

Although all of these paths are probably tested indirectly via
test model parsing in various other test cases, the concrete attribute
parsing is not directly tested anywhere. This makes the expected output
of the core parsing logic a bit harder to reason through.

This is the first of three commits to add direct tests for the core parser
logic, which will make updates and changes safer and easier to reason about.
This commit adds test coverage for metric parsing. All tests have YAML
inline instead of saved in a file, as the file and directory reading logic
is outside the scope of the metric parser itself and the metric config spec
is simple enough to comfortably fit inside of each test function.