Add tests for materialization config parsing
Second of three commits adding test coverage on core parsing logic.
This one covers all of the materialization object types, especially
the one object implementing custom string input parsing.

Also of note for materializations is the nested list construct for
the `rollups` attribute, which presents something of an edge case.

Please see the recent commit titled "Add tests for metric config parsing"
for more background on the reasoning behind this set of changes.