Add tests for data source config parsing
Third of three, this adds test coverage for bare data source parsing
with various options enabled. There are no error states to check at
the moment because data source properties are all fairly simple.
Note this includes a base test for the metadata parsing for the measure
objects as well, since those are defined in the datasource and have
metadata propagated into them as part of that parsing process.

See commit titled "Add tests for metric config parsing" for more context
on this change.