Fix custom object parsing for internal model initialization
If we initialize a model internally, the Pydantic validation will
still invoke the parsing and validation methods provided via the
`__get_validators__` magic method. This means passing in an instance
of a model object using our PydanticCustomInputParser interface will
result in a runtime validation error.

Since this is silly, this commit updates all of those custom parse
methods to pass an instance of the object through. We cannot do this
in the base class, because you can't assert a value to be an instance
of the generic return type. This is probably for the best, as we
might well want to only provide this instance pass through for some
subset of these classes. For now we update them all, and fix up
their docstrings along the way.