Switch metric measure references from string name to model object
Our Metric model object takes in MetricTypeParams that allows for various
input measure name references. These are all typed as strings or
lists of strings, since the current MetricFlow framework only requires
a pointer to the original measure model object.

Upcoming changes will add constraint expressions to the measure
pointers stored inside of the Metric model object, and as such we
will need to attach additional properties to each of the measures
referenced in the MetricTypeParams.

This commit simply introduces a new model object type to represent the
measure inputs for a metric, and converts the string types to use the
new object. All other changes are simply to manage the fallout of the
related type change. Subsequent commits will update the jsonschema and
enable user input of object-type representations and add new properties
as appropriate.