Enable YAML object parsing for metric input measures
Now that the internals of the framework rely on the MetricInputMeasure
objects, we can allow users to specify either the legacy string name
representation of the measure or the full object representation.

This sets us up to add properties, such as constraints, to the
measures defined as inputs to metrics in the model.