Centralize YAML loading context-aware helpers
Metricflow's YAML file loading uses two different loader types,
depending on whether we want file parsing context metadata included
in the payload or not. These were invoked via the one-line
yaml.load_all calls with the relevant parser (although sometimes we'd
use the wrong one and load context data in only to remove it later).

This causes problems with our context-aware parsing, because PyYAML
will only correctly populate the `name` field on its output nodes if
the input stream is a file-like object with the `name` property set
correctly.

This commit sets us up for streamlining the parser by centralizing
the logic for converting an input string into a file-like object
and requesting the context-aware loader when needed. In order to
keep things a bit more readable all callsites in the core parser
and test parser were migrated to the new helper methods, while the
CLI config parsing was left as-is since it is already in a custom
helper class of its own.