Source YAML contents for metadata from parsing context
The existing parser sources YAML contents for a given node by passing
the complete YAML as a list of lines through the entire parser stack,
and extracting segments based on the line number values attached to
the parsed output in the ParsingContext object.

This is unnecessary, and also prevents us from relying on things like
Pydantic's parsing abstractions for populating the metadata field
in a more re-usable manner. This commit eliminates the threading of
content values through the stack by instead sourcing them from the
YAML node available to the custom loader when the ParsingContext
is populated. As documented, the ParsingContext holds the node, rather
than a serialized YAML string, because we'd rather serialize lazily
instead of re-serializing every node at every level regardless of
whether or not the metadata object will be populated.