Merge the logic of 'validate_config_structure' into 'parse_config_yaml'
The method of 'validate_config_structure' was duplicating a lot of
(almost all of) the logic of 'parse_config_yaml'. This commit merges
the "extra" stuff 'validate_config_structure' was doing (validating
the yaml conformed to the jsonschema spec). This in turn means that:
we load the yaml documents one less time, we iterate over the yaml
documents one less time, and we reduce the number of duplicate errors
being produced, and reduce code complexity by keeping things more
DRY.