Build a 'ModelValidationResults' during parsing which is optionally returned
Instead of collecting and raising errors in 'parse_directory_of_yaml_files_to_model'
and the methods it calls, we now collect 'ValidationIssues' and build a
'ModelValidationResults' object. The method also takes a new argument
'raise_issues_as_exceptions'. If 'raise_issues_as_exceptions' is 'True', then
the default behavior of the method is the same, in that if there are any 'ERROR'
level 'ValidationIssues' it raises an exception with the message being a
concatenation of all the issues. However, if 'raise_issues_as_exceptions' is
'False' then even if there are 'ERROR' level issues, a ModelValidationResults
object is returned and it is expected the caller will handle the results.