Make UserConfiguredModel required for ModelBuildResult
Over the past month it has become the case that ModelBuildResult is
never instantiated without a UserConfiguredModel. We want to continue
this pattern. To ensure this, we should make the model property of
ModelBuildResult required, which this commit does. This also makes
it so that we no longer have to assert that a ModelBuildResult has
a model.