Better exception handling flow for 'parse_config_yaml'
Previously we weren't exception handling on a document by document
basis. Which means if there was an issue in one document in the config
yaml, we'd cease validating the rest of the documents. This wasn't
desireable, because it means if there are multiple issues, you'd only
find out about the first one run into each time parsing is run. This
in turn makes fixing issues in configs slow and tedious. Now in, in
most cases, if there is an exception with a document, we catch it early
and continue validating/parsing the rest of the documents.