Process constraint specifications on distinct measures
This hooks up the measure constraints to the DataflowPlanBuilder,
allowing us to process the constraint and render SQL for propagating
the constrained measure value up to the metrics computation part
of the process. With this commit users can now specify a constrained
measure as input to a metric and get the correct output.

Note this does NOT work for constraints applied to the same measure.
Doing so requires a fair bit of work on propagating column aliases,
and as such it will be handled in a separate commit.