Remove 'load_all_without_context'
Previously we'd been using 'load_all_without_context', to load config
yaml for jsonschema validation. However, now we use 'load_all_with_context'
in all cases, and want to encourage this as a continued behavior. This is
because 'load_all_with_context' adds a '__parsing_context__' to parsed yaml
and if someone were to specify a '__parsing_context__' in their yaml, it
will be clobbered by 'load_all_with_context', which is desired.