Add FileParsingResult class to enhance clarity of what 'parse_config_yaml' returns
The method 'parse_config_yaml' was returning a tuple of two lists. This was
confusing because there exists a convention wherein when a tuple of two lists
is returned, it is assumed that the lists are a one to one mapping, i.e. that
element x of the first list is related to elment x of the second list. This
convention of association was not intended for the return of
'parse_config_yaml', thus we added a new dataclass 'FileParsingResult' to
clarify what is being returned and avoid this unintentional assumed
association.