Add support for aliased input measures
With the advent of measure-specific constraints defined as inputs to
metrics, we encountered a scenario where it's possible for a user to
pass the same measure - with different constraints - to a single
metric twice. This results in problems with our internal SQL
rendering, which will do strange things with the seemingly
duplciated column name.

While it might be possible to automate an aliased solution internally
for ratio metrics, for expr-type metrics this is impossible. Ultimately,
the end user will need to be able to reference the constrained measures
in the metric expr entry in their configs. The only way for them to be
able to reliably do this is for Metricflow to provide the option of
giving input measures an alias for use as a reference in expr entries.

This commit adds support for this construct along with some additional
tests illustrating the process.

One key point with the approach here is that it is a bit brittle - the
surgical insertion of the aliasing logic into the aggregate measures node,
while convenient, leaves us with no guarantees that we won't encounter
problems as we add more downstream consumers of this node. However,
based on the code as I understand it there is no viable path to providing
strong guarantees in any other way without either plumbing alias information
through the entire dataflow plan (and corresponding SQL plan) or jettisoning
the visitor pattern in favor of something more strictly typed in terms
of node-level input dependencies. While these might be worth pursuing later,
at this stage such a level of restructuring did not seem warranted.