Disallow metric_time as an element name
With the advent of the metric_time query construct it has effectively
become a reserved keyword for Metricflow, as we cannot resolve queries
that are requesting `metric_time` if there is some other time dimension
also named `metric_time` in the user model.

This commit adds a general MetricFlowReservedKeywords enumeration and
adds a validation check to assert that none of the checked names
match a value in the enum.