Add readme and other metadata info to pyproject.toml
Our pypi page is essentially blank, and it appears this is because
we have not filled out the pyproject.toml fields poetry requires
to push any of our useful information to pypi.

With other package managers it's possible to customize the pypi
description to render, e.g., a brief header followed by the
changelog. Poetry does not allow this, and instead requires a pointer
to a README.md or some other specific readme location/format combo,
so as a comprise the changelog is now linked from the top of the
README.md as well.