Refactor data warehouse validations to use 'sql_query_plan_renderer' to render queries
This change is two fold. First generating queries for measures and identifiers
is fair bit harder than the quries we'd been generating for data_sources and
dimensions, which was going to increase the complexity of the task builder
greatly. So we wanted to begin using the tools the rest of MetricFlow uses
for generating queries, no need to reinvent the wheel. Additionally, by using
the 'sql_query_plan_renderer' we ensure validation queries are similar to what
MetricFlow would render using the same elements.

TLDR: This makes future work on data warehouse validations easier and makes
validations consisten with the rest of MetricFlow.