BUG: fix stragglers missed from column reordering
There are cases where the visitor just adds the metric_select_columns and the non_metric_select_column_set.as_tuple(). This leads to the ordering being this (metric, others…). Since the ordering is sorted in the as_tuple() method. So in order to make the ordering the way we wanted it to be, we have to combine the columns sets and use as_tuple on the entire thing.