Ignore specs with identifier_links for data warehouse validations
The spec tuples created by 'create_sql_source_data_set' include the base
specs not including any 'identifier_links' as well as those same
specs but with 'identifier_links'. For the purpose of data warehouse
validations, the latter are unnecessary as we're only trying to verify
that the thing listed exists.