Remove fatal validation issues as a concept
Fatal validation issues comes from a time when our validators weren't
resilient enough, wherein one thing being bad meant other validations
couldn't be run properly. This is no longer the case, it hasn't been
for awhile. Fatal issues are actually problematic because they can
slow down the cycle of writing configs. This is because fatals mean that
instead of getting all the semantic validation issues, you only get
issues up until a fatal issue is reached. In our current process,
parts of validation are run in their entirety in order: linting ->
parsing -> semantic -> data warehouse. If there is an error in a
a given stage, the next stage cannot be entered. But, in this system
errors are the gate keepers. Fatals are no more.