Multi-thread semantic validations
This might be a premature improvement, but it is a matter of time before
it becomes absolutely necessary. Semantic validations are CPU bound. They
dont depend on any external information. As we add more semantic validation
rules and as configs grow larger, semantic valdiations become slower.
Additionally, semantic validations don't depend on one another and they
don't alter the underlying model. Thus the straight forward solution is to
allow more CPU threads to be thrown at them. On a small config, there is
virtually no difference in runtime currently. On exceedingly  large configs,
time gains are seen (and to different degrees based on how many threads
there are).