Enforce a 1:1 match between aliased measures and input measures
When measures are constrained and and aliased, we apply the
alias to the instance set of measures immediately following the
aggregation step. Since each aggregation should have exactly one
measure as input, there should always be a 1:1 match between an
aliased measure output and a measure input.

If we ever push an optimization or other change that breaks this
assumption we run the risk of aliasing things in a way that
results in incorrect metric computation.

Longer term this should all be merged into the MeasureSpec, which
would naturally enforce a 1:1 alignment between a given measure
computation and an associated alias, or else we should find some
other way of preventing this issue from arising.