Increase strictness of mypy typechecking
Metricflow originally used the stock mypy options for typechecking,
but recent PRs involving more direct worth with Pydantic and
jsonschema objects and methods (which tend to be Any or incompletely
typed due to the nature of JSON values) has resulted in a
proliferation of incompletely typed methods.

This commit adds a mypy.ini with a few initial options in place to
rein in potential accidental spread of Any types and incompletely typed
methods, and fixes the relevant mypy errors. We do this in one
fell swoop just because the total number of errors was fairly small.