Ensure warehouse validation tables exist when needed
Pytest fixtures only get invoked when they are included in the fixture,
call graph, so the session-scoped fixture was getting called once in
this test module. This worked out fine as long as the test case
referencing the fixture was invoked before the test cases that
needed it, but that isn't guaranteed to happen when parallel
execution gets involved.

This should resolve the issue by ensuring the fixture is invoked when
needed.