Add concrete Snowflake implementation of data warehouse inference context provider
This commit adds a concrete implementation for fetching data warehouse context from Snowflake. This is only temporary, since the ideal solution would be to use the existing `SqlEngine` classes to do that. However, since we want to test inference without tampering with the core engine and adding features to it, this has been implemented in a separate class.

As data source inference becomes a stable feature, we should definitely make it use Metricflow's core, and add sampling to it.