Add base class for inference solver
This commit adds the base class for inference solvers. An inference solver is a policy/algorithm that takes as input a bunch of signals that were produced by inference rules. It is then responsible for producing a final verdict about what a column is or isn't, or indicate any problems that might have happened during the process.