Add inference solver that uses a weighted type tree
This commit adds the `WeightedTypeTreeInferenceSolver`. It assigns weights to each node in the type tree, and traverses the path with the most weight until it finds a leaf, the weight trail stops or it is confused, i.e there is a bifurcation in the path with the most weight according to some threshold.