Add `InferenceRunner` class with progress reporting
This commit adds an `InferenceRunner` class that acts as a "main" function for inference. You can set up context providers, rules, solvers and renderers and make it run them as a complete pipeline.

It also supports a `InferenceProgressReporter` for reporting progress. This can be used for checking on the status of inference (which may take several seconds) in the CLI or in a script.