[SCD] Add validity_info dimension property to config schema
To designate a data source as containing versioned dimension values,
such as one might associate with an SCD (slowly changing dimension)
Type II database table, we need a mechanism for indicating that a
given data source contains dimensions, and how to determine which
value set is valid (i.e., which row contains the values we need
for the given dimension key).

This adds parameters allowing users to specify a given (time) dimension
as the start or end of a validity window, which will in turn enable a
join filter to compare the measure agg time against for our dimension
joins. Subsequent changes will implement processing and add validation
to enforce only supported configurations can be used.