Add row with non-existent listing to fct_bookings test data
Our test datasets are incredibly well-behaved.

This change makes them a little more realistic by adding an entry
with an invalid listing ID to the fct_bookings data source, which
increases our confidence in our test cases and allows us to evaluate
how our joins handle missing values.

This also fixes up a test case which was improperly configured
to use an INNER JOIN when the framework is expecting a LEFT JOIN.