Add dimension inference rule for low cardinality strings
User reports show that we are missing a lot of "obvious" dimension
classifications in inference where we are unable to detect the type
of the column.

The root of the problem appears to be that our STRING column type
match rule only applies if we already know the base type to be
DIMENSION, but otherwise we skip it.

Increasing the scope of that rule to tag all STRING type columns
as dimensions seemed overly broad, since UUIDs are a thing, but
having a matcher rule that checks for both STRING and relatively low
cardinality seems to be a nice improvement.

This was checked against an internal snowflake table and produced
the desired results - a previously unmatched field was tagged as
a dimension.

Note the foreign key match rule cardinality was bumped up a touch
as well, since we can generally accept a higher range of cardinalities
to say that a given identifier is a foreign key.