Move WhereClauseConstraint conversion out of the query parser
We occasionally need to convert WhereClauseConstraint objects to
SpecWhereClauseConstraint objects. For historical reasons
this logic was held in the MetricFlowQueryParser, but keeping
it there does not make a lot of organizational sense. Worse, it
sets up potential circular dependencies, as it forces the query
parser to depend on our semantic containers, and the dataflow
plan builder to depend on the query parser.

Here we split this helper into its own independent class. We cannot
move it onto either the source or target classes because of other
concerns around possible circular dependencies, which we have
documented in the new module.