Move AggregationType to a root module
The AggregationType enum was included with the Measure model object
because that was where it was first needed, but its utility is quite
a bit broader than as a model enumeration. Since its current location
leads to potential circular imports, and additional aggregation
properties (such as additiveness) might be coming in the future, we
move it to a root aggregation_properties module.