Improve error message for missing aggregation time dimension
If the user fails to include either a primary time dimension or a
measure-specific aggregation time dimension for a measure in their
model they will encounter an assertion check either during validation
or querying. The original error message was really intended for
developers, but it is incidentally raised in validation. This quick fix
makes it a bit more user-friendly.