Handle exception during building SemanticModel in materializations validation.
Many things can cause building of the SemanticModel from the UserConfiguredModel
to blow up. Noteably, if a dimension/measure referenced by a metric doesn't exist.
This has caused the Materializations Validation to fail, even if the issue is
unrelated to materializations. Additionally, a very arcane error message was being
output because part of the issue raised includes the arguments to the called
function, which in this case was the entire UserConfiguredModel.

This commits aims to fix the problem by handling the error and letting the user
know what went wrong in a friendly manner.