Tighter scoping of exceptions caught during materializations validation setup
It was brought up that perhaps we shouldn't catch ALL exceptions in the
materialization validation rule setup when the semantic model is built.
This scopes the exceptions handled to AssertionErrors and SemanticExceptions.
These two types cover likely all the exceptions taht we "expect" to happen
when building a SemanticModel for a UserConfiguredModel.