Refactor dataclass serializer to use field definition specs
The original implementation of the dataclass serializer passed
type specifications around, probably due to the lack of reliable
type availability in the dataclasses module APIs. Unfortunately,
this makes it impossible to handle default values correctly.

As preparation for wiring proper defaults through, we restructure
the code slightly to use field definition objects that can carry
type and default value information as needed.