Ensure measure constraints and semi-additive measures work together
There are potential boundary cases in play when applying constraints
to semi-additive measures. This provides some test coverage for that
interaction.

Note the test case itself encodes an expectation based on current behavior
of semi-additive measures, which do not return the full set of
possible rows if the user groups by the restricted dimension. Later
updates can tweak this test case to handle different constraint
handling as needed.