Add tests for measure constraints with a single aliased measure
In most cases with measure constraints there will be more than
one measure (ratio metrics, multi-measure expr metrics), and we can
rely on the existing join -> filter path to ensure proper use of
the alias. However, the single measure case runs through a separate
code path in the dataflow plan builder.

This works today because the alias is internal to the aggregated
measures node, and the expr is required to use the alias rather
than the name, so the aggregated measures node gets an alias applied
and the query proceeds as expected.

This case was not tested previously, but it's sufficiently important
and easily overlooked, so we add a case to cover it.