Replace Identifier class init method with pydantic validator
Since MetricFlow has been public, the Indentifier class has used an __init__
method for setting some defaults for instatiated instances. The problem with
this is that by using __init__ the pydantic validators get clobbered. Thus
this commit moves the __init__ attribute setting into pydantic validators.
Specifically a validator for setting the default of 'entity'.

Additionally, you may notice that previously we were setting the default of
'identifiers' in the __init__. This was because we define 'identifiers' as
'identifiers: List[CompositeSubIdentifier] = []' which lets python typing know
that 'identifiers' will never be 'None', but had the side effect of all instances
with the attribute defined default to share a referance to a single list. Thus
we had to set the 'identifiers' default int the __init__ to ensure each instance
had it's own list. However, this this secondary defaulting is no longer necessary
because pydantics base validators handle this.