Add validation for constrained measure aliases
The addition of measure constraints opens up the possibility of
two different variations on the same measure. When this happens users
must provide an alias for each distinctly constrained variation of
the input measure.

This adds a validator to enforce that model contract, and to provide
better user-facing error messages than whatever inscrutable SQL
errors might arise as a result of rendering two different measure
inputs with the same alias. We also enforce our standard naming
constraints, including that aliases must not collide with measure
names. This last is potentially unnecessary, but until we get more
experience with our query layouts and address possible downstream
optimizations it is prudent to apply a more restrictive policy.