Move metric time instance extraction into helper
There are a few places where we do the same SqlDataSet -> collection
of metric time instance operation, and another is coming soon. This
moves the core operation into a helper, and along the way makes
it more obvious that the metric time reference object in the
JoinOverTimeRangeNode is the same one provided by the central
DataSet.get_metric_time_reference() method.