Propagate validity window metadata into join resolution logic
The validity window metadata is stored in the model, and only really
used when doing join resolution to configure the ON condition for the
joins themselves. Furthermore, we only need the validity window in
the specific case where we are attempting to process joins against
SCD data sources.

This commit adds accessors for fetching the validity window out of
a DataSourceSemanticsAccessor and propagating it as part of the
join description at join resolution time. Now that this information
is available in the resolved nodes in the DataflowPlan, we can
make use of it for rendering the relevant join conditions. For now,
we simply pass the metadata along and ensure it is available.