Add test for dataflow to sql plan conversion for SCDs
This adds the test and snapshot files for the plan conversion. Since
the SQL rendering for the SCD condition is not yet implemented, the
ON conditions are incorrect. A follow-up commit will add the
rendering support and update the plans with the relevant changes.

Ordinarily we'd like to have the queries be complete and correct,
but in this case the change is sufficiently subtle that we split
things up to highlight the difference.