Convert validity window join conditions in sql plan generation
An SCD data set with a validity window requires a time range filter
in the ON condition in order to produce correct results - without it
we would be including invalid rows and skewing the aggregated result
output.

This commit updates the join condition to account for validity windows
in the conversion from dataflow to sql plan, thereby rendering the
relevant comparison check in our output SQL. This works for both
single-hop and multi-hop cases, although it only does so because we
do not currently allow for fanout joins. It will also fail on a
boundary case where users attempt a join that uses two SCD style
data sources. Subsequent commits will deal with these boundary cases,
and ensure that anyone enabling fanout will need to explicitly handle
the SCD case when doing so.