Intial dbt manifest to user_configured_model transformation
This is somewhat of an unstructured commit. It was initially a checkpoint
commit and is now hard to break up. It sets up the process for iterating
over dbt defined metrics found on the dbt manifest and turning them into
metricflow data sources, data source sub elements, and metrics. There is a
lot this commit doesn't do though like: building metrics, building measures,
getting into the complexity of time dimensions, and etc.

Additionally, this commit is noticeably devoid of method docstrings. This
will be resolved in the future commits (and should be included in the same
PR). Apologies that this makes it slightly harder to grok what is going on.